challange api call and login redirect fix
