+++ package.json
@@ -55,6 +55,7 @@
     "@types/react-native": "^0.72.8",
     "eslint": "^9.25.0",
     "eslint-config-expo": "~9.2.0",
+    "reactotron-react-native": "^5.1.13",
     "typescript": "~5.8.3"
   },
   "private": true
